Remove unused train
