--- satflow/experiments/pl_train.py
@@ -1,98 +1,0 @@
-from typing import List, Optional
-
-import hydra
-from omegaconf import DictConfig
-from pytorch_lightning import (
-    Callback,
-    LightningModule,
-    LightningDataModule,
-    Trainer,
-    seed_everything,
-)
-from pytorch_lightning.loggers import LightningLoggerBase
-from pytorch_lightning.profiler import AdvancedProfiler, PyTorchProfiler
-from satflow.core import utils
-from satflow.core.callbacks import NeptuneModelLogger
-from pytorch_lightning.callbacks import LearningRateMonitor
-
-log = utils.get_logger(__name__)
-
-
-def train(config: DictConfig) -> Optional[float]:
-    """Contains training pipeline.
-    Instantiates all PyTorch Lightning objects from config.
-
-    Args:
-        config (DictConfig): Configuration composed by Hydra.
-
-    Returns:
-        Optional[float]: Metric score for hyperparameter optimization.
-    """
-
-    # Set seed for random number generators in pytorch, numpy and python.random
-    if "seed" in config:
-        seed_everything(config.seed, workers=True)
-
-    # If required
-    # Init Dataloaders
-    log.info(f"Instantiating datamodule <{config.datamodule._target_}>")
-    datamodule: LightningDataModule = hydra.utils.instantiate(
-        config.datamodule, _convert_="partial"
-    )
-
-    # Init Lightning model
-    log.info(f"Instantiating model <{config.model._target_}>")
-    model: LightningModule = hydra.utils.instantiate(config.model)
-
-    # Init Lightning callbacks
-    lr_monitor = LearningRateMonitor(logging_interval="step")
-    callbacks: List[Callback] = [lr_monitor, NeptuneModelLogger()]
-    if "callbacks" in config:
-        for _, cb_conf in config["callbacks"].items():
-            if "_target_" in cb_conf:
-                log.info(f"Instantiating callback <{cb_conf._target_}>")
-                callbacks.append(hydra.utils.instantiate(cb_conf))
-
-    # Init Lightning loggers
-    logger: List[LightningLoggerBase] = []
-    if "logger" in config:
-        for _, lg_conf in config["logger"].items():
-            if "_target_" in lg_conf:
-                log.info(f"Instantiating logger <{lg_conf._target_}>")
-                logger.append(hydra.utils.instantiate(lg_conf))
-
-    # Init Lightning trainer
-    log.info(f"Instantiating trainer <{config.trainer._target_}>")
-    trainer: Trainer = hydra.utils.instantiate(
-        config.trainer,
-        callbacks=callbacks,
-        logger=logger,
-    )
-
-    # Send some parameters from config to all lightning loggers
-    log.info("Logging hyperparameters!")
-    utils.log_hyperparameters(
-        config=config,
-        model=model,
-        trainer=trainer,
-    )
-
-    # Train the model
-    if config.trainer.auto_lr_find or config.trainer.auto_scale_batch_size:
-        log.info("Starting tuning!")
-        trainer.tune(model=model, datamodule=datamodule)
-    log.info("Starting training!")
-    trainer.fit(model=model, datamodule=datamodule)
-
-    # Evaluate model on test set after training
-    if not config.trainer.get("fast_dev_run", False):
-        log.info("Starting testing!")
-        trainer.test(model=model, datamodule=datamodule)
-
-    # Print path to best checkpoint
-    log.info(f"Best checkpoint path:\n{trainer.checkpoint_callback.best_model_path}")
-
-    # Return metric score for hyperparameter optimization
-    optimized_metric = config.get("optimized_metric")
-    if optimized_metric:
-        return trainer.callback_metrics[optimized_metric]
